Add module for generically handling file storage
The `channel.storage` module defines a `Storable` protocol for
defining various ways of storing file data. In this case only
one implementation, `FileSystemStorage` is provided, which as
the name implies works on the local filesystem, but in the future
it could be possible to add something like AWS S3 storage.

Tests are added to validate this functionality, testing both the
underlying utilities (E.g. `path-relative-to-root`), but also for
testing the operation of file system storage itself.