Update documentation to refer to objects not files
The `Storage` protocol is for storing any type of object, not just
files. A better way to display this is to regard it as storage of
generic objects, each implementation defining what the definition of
that object is. In the filesystem case an object is a file.

The protocol still expects data streams to be used however, so
serialization is handled by the caller.